Improve code readability: extract repeated code and func
