+++ lib/helpertools/helpertools/__init__.py
@@ -1,2 +1,3 @@
+from helpertools.geometry import *
 from helpertools.lambda_helpers import *
 from helpertools.pdf_processor import *
