+++ selectable_pdf_infra/lambda_layer/helpertools_py38/createlayer.sh
@@ -12,6 +12,7 @@ if [ "$1" != "" ] || [$# -gt 1]; then
 	rm  -rf $lib_name
 	# ln -s $lib_dir $lib_name  #cannot use sym link in docker
 	cp -r $lib_dir .
+	rm -rf $lib_name/build $lib_name/dist $lib_name/textracttools.egg-info
 	docker run -v "$PWD":/var/task "lambci/lambda:build-python$1" /bin/sh -c "pip install $lib_name/ -t python/lib/python$1/site-packages/ --upgrade; exit"
 	zip -r $lib_name.zip python > /dev/null
 	rm -rf $lib_name
